conect roomsGrid to data base and not local
