+++ psagot-client/src/pages/HomePage.jsx
@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { fetchAllUserTypes } from "../features/userType/userTypeActions";
 import { Typography, Button, Container } from "@mui/material";
-import RoomsGrid from "./RoomsGrid";
+
 
 const HomePage = () => {
     const dispatch = useDispatch();
@@ -26,7 +26,6 @@ const HomePage = () => {
         <Container item style={{ textAlign: 'center', padding: 10 }}>
             <Typography variant="h5">😀hello psagot project😀</Typography>
             <Button onClick={handleClickButton}>Example of a function structure</Button>
-            <RoomsGrid/>
         </Container>
     );
 }
