+++ psagot-client/src/routers/AppRouter.js
@@ -2,12 +2,13 @@ import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import HomePage from '../pages/HomePage';
 import Layout from '../pages/Layout';
+import RoomsGrid from '../components/RoomsGrid';
 
 const AppRouter = () => {
     return (
         <Routes>
             <Route path='/' element={<Layout/>}>
-                <Route path='/' element={<HomePage/>}/>
+                <Route path='/' element={<RoomsGrid/>}/>
             </Route>
         </Routes>
     );
